feat: implement linkification utility and integrate into ClickableLinksPlugin
